Specgram thresholding
Implemented spectrogram thresholding via different methods:
Global methods: kmeans, median, isodata, yen and otsu methods
Adaptive methods: gaussian, mean, median (skimage.filter)

 Changes to be committed:
	modified:   lib/a2audio/recanalizer.py
	modified:   lib/a2audio/roizer.py
	new file:   lib/a2audio/thresholder.py
	modified:   scripts/PatternMatching/audiomapper/align.py
	modified:   scripts/PatternMatching/audiomapper/recnilize.py

+++ scripts/PatternMatching/audiomapper/align.py
@@ -23,7 +23,7 @@ for line in sys.stdin:
     columns = int(columns.strip('\n'))
     jId = int(jobId.strip('\n'))
 
-    spec = numpy.zeros(shape=(0,int(columns)))
+    spec = numpy.zeros(shape=(0,int(columns))).astype('int')
     for s in spectro:
         row = s.split(',')
         row = map(float,row)
